restructured file layout and refactored the simplebar code
