+++ src/routes/(authenticated)/(dashboard)/_dashboard.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, Outlet } from "@tanstack/react-router";
-import VerticalNav from "../../../components/vertical-nav";
+import VerticalNav from "../../../components/layout/vertical-nav";
 
 export const Route = createFileRoute("/(authenticated)/(dashboard)/_dashboard")(
   {
@@ -14,7 +14,7 @@ function Dashboard() {
         <div>
           <VerticalNav></VerticalNav>
         </div>
-        <div className=" min-h-screen p-10">
+        <div className="min-h-screen ml-[300px] p-10">
           <Outlet />
         </div>
       </div>
